fix some compiler warnings

+++ dk.h
@@ -53,7 +53,7 @@ bool DK_on_char(SC ch);
 void DK_print_key_names();
 
 int DK_name_to_index(char *name);
-char *DK_index_to_name(int i);
+char *DK_index_to_name(size_t i);
 
 DK_Node *DK_pivotNode_fetch(DK_Node * const node, DK_Key key, bool pivot);
 bool DK_descend(DK_Key key);
